requirements.txt
django~=4.0.0
uwsgi
yt-dlp
youtube-dl
pillow
django-filter
djangorestframework
markdown
numpy
pandas
py-deezer
pyacrcloud
requests

